mon travail integré avec wassim

+++ src/Controller/RendezVousController.php
@@ -8,25 +8,23 @@ use App\Entity\User; // On utilise User à la place de Patient
 use App\Form\RendezVousType;
 use App\Repository\UserRepository; // Import du UserRepository
 use App\Repository\DisponibiliteRepository;
-use App\Repository\MedecinRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use App\Service\AIService;
 
 class RendezVousController extends AbstractController
 {
 #[Route('/rendezvous/dispo/{id}', name: 'app_rendezvous_dispo')]
-public function afficherDispos(int $id, MedecinRepository $medecinRepo, DisponibiliteRepository $dispoRepo): Response 
+public function afficherDispos(int $id, UserRepository $medecinRepo, DisponibiliteRepository $dispoRepo): Response 
 {
     $medecin = $medecinRepo->find($id);
 
+
     $dispos = $dispoRepo->findBy(
-        [
-            'medecin' => $medecin, 
-            'est_reserve' => false  
-        ],
+        ['medecin' => $medecin],
         ['date_debut' => 'ASC']
     );
 
@@ -36,50 +34,52 @@ public function afficherDispos(int $id, MedecinRepository $medecinRepo, Disponib
     ]);
 }
 
-    #[Route('/rendezvous/reserver/{id}', name: 'app_rendezvous_reserver')]
-    public function reserver(Disponibilite $dispo, Request $request, EntityManagerInterface $em): Response
-    {
-        // 1. Vérification de sécurité
-        if ($dispo->isEstReserve()) {
-            $this->addFlash('danger', 'Ce créneau est déjà réservé.');
-            return $this->redirectToRoute('app_rendezvous_dispo', ['id' => $dispo->getMedecin()->getId()]);
-        }
-
-        $rdv = new RendezVous();
-        
-        // 2. Pré-remplissage automatique
-        $rdv->setMedecin($dispo->getMedecin());
-        $rdv->setDatetime($dispo->getDateDebut());
-        $rdv->setDisponibilite($dispo);
-        $rdv->setStatut('EN_ATTENTE');
-        
-        
-        // On lie le patient connecté
-        $user = $this->getUser();
-        if ($user) {
-            $rdv->setPatient($user); // On passe l'objet User directement
-        }
-
-        $form = $this->createForm(RendezVousType::class, $rdv);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            // 3. Logique de confirmation
-            $dispo->setEstReserve(true);
-            $em->persist($rdv);
-            $em->flush();
-
-            $this->addFlash('success', 'Demande de rendez-vous envoyée au Dr. ' . $rdv->getMedecin()->getUser()->getNom() . '. Vous serez notifié dès qu\'il aura accepté ou refusé.');
-            return $this->redirectToRoute('app_medecin_recherche'); 
-        }
-
-        return $this->render('rendezvous/finaliser.html.twig', [
-            'form' => $form->createView(),
-            'dispo' => $dispo
-        ]);
+#[Route('/rendezvous/reserver/{id}', name: 'app_rendezvous_reserver')]
+public function reserver(Disponibilite $dispo, Request $request, EntityManagerInterface $em, AIService $aiService): Response 
+{
+    // 1. SÉCURITÉ : Si un patient accède à l'URL d'un créneau déjà réservé entre-temps
+    if ($dispo->isEstReserve()) {
+        $this->addFlash('danger', 'Désolé, ce créneau n\'est plus disponible.');
+        return $this->redirectToRoute('app_medecin_recherche');
+    }
+    
+    $rdv = new RendezVous();
+    $rdv->setMedecin($dispo->getMedecin());
+    $rdv->setDatetime($dispo->getDateDebut());
+    $rdv->setDisponibilite($dispo);
+    $rdv->setStatut('EN_ATTENTE');
+    
+    if ($user = $this->getUser()) { 
+        $rdv->setPatient($user); 
     }
 
-    #[Route('/medecin/rendezvous/{id}/accepter', name: 'app_medecin_rdv_accepter')]
+    $form = $this->createForm(RendezVousType::class, $rdv);
+    $form->handleRequest($request);
+
+    if ($form->isSubmitted() && $form->isValid()) {
+        $motif = $rdv->getMotif();
+        $rdv->setPriorite($motif ? $aiService->calculerPriorite($motif) : 2);
+
+        // 2. IMPORTANT : On marque la disponibilité comme réservée
+        // Cela va la faire disparaître du calendrier grâce au filtre Twig
+        $dispo->setEstReserve(true);
+        
+        $em->persist($rdv);
+        // persist($dispo) n'est pas obligatoire car Doctrine surveille déjà l'objet
+        
+        $em->flush();
+
+        $this->addFlash('success', 'Votre demande a été envoyée. Ce créneau vous est maintenant réservé en attente de validation.');
+        return $this->redirectToRoute('app_medecin_recherche'); 
+    }
+
+    return $this->render('rendezvous/finaliser.html.twig', [
+        'form' => $form->createView(),
+        'dispo' => $dispo
+    ]);
+}
+
+#[Route('/medecin/rendezvous/{id}/accepter', name: 'app_medecin_rdv_accepter')]
 public function accepter(RendezVous $rdv, EntityManagerInterface $em): Response
 {
     $rdv->setStatut('CONFIRME');
@@ -107,27 +107,25 @@ public function refuser(RendezVous $rdv, EntityManagerInterface $em): Response
 
 
 
-    #[Route('/rendezvous/annuler/{id}', name: 'app_rendezvous_annuler', methods: ['POST', 'GET'])]
+
+#[Route('/rendezvous/annuler/{id}', name: 'app_rendezvous_annuler', methods: ['POST', 'GET'])]
 public function annuler(RendezVous $rdv, EntityManagerInterface $em): Response
 {
-    // 1. Récupérer la disponibilité liée à ce rendez-vous
-    $dispo = $rdv->getDisponibilite();
-
-    if ($dispo) {
-        // 2. Remettre le créneau comme disponible pour les autres patients
-        $dispo->setEstReserve(false);
+    // Vérification de sécurité : seul le patient propriétaire peut annuler
+    if ($rdv->getPatient() !== $this->getUser()) {
+        throw $this->createAccessDeniedException("Vous ne pouvez pas annuler ce rendez-vous.");
     }
 
-    // 3. Supprimer le rendez-vous de la base de données
-    // Note: Vous pouvez aussi faire $rdv->setStatut('ANNULE') si vous voulez garder une trace
+    $medecinId = $rdv->getMedecin()->getId();
+
+    // On supprime simplement la demande
     $em->remove($rdv);
-    
-    // 4. Sauvegarder les changements
     $em->flush();
 
-    $this->addFlash('success', 'Votre rendez-vous a bien été annulé. Le créneau est à nouveau libre.');
+    $this->addFlash('success', 'Votre demande de rendez-vous a été annulée.');
 
-    return $this->redirectToRoute('app_home'); // Ou vers la page "Mes Rendez-vous"
+    // On redirige vers la page du calendrier du médecin pour voir le bouton "Réserver" revenir
+    return $this->redirectToRoute('app_rendezvous_dispo', ['id' => $medecinId]);
 }
